+++ bin/cli.js
@@ -3,7 +3,7 @@
 var subcommand = require('subcommand')
 var debug = require('debug')('dat')
 var usage = require('../src/usage')
-var parseArgs = require('../src/parse-args')
+var pkg = require('../package.json')
 
 process.title = 'dat'
 
@@ -12,17 +12,35 @@ var NODE_VERSION_SUPPORTED = 4
 var nodeMajorVer = process.version.match(/^v([0-9]+)\./)[1]
 var invalidNode = nodeMajorVer < NODE_VERSION_SUPPORTED
 if (invalidNode) exitInvalidNode()
+else {
+  var notifier = require('update-notifier')
+  notifier({ pkg: pkg })
+    .notify({
+      defer: true,
+      isGlobal: true,
+      boxenOpts: {
+        align: 'left',
+        borderColor: 'green',
+        borderStyle: 'classic',
+        padding: 1,
+        margin: { top: 1, bottom: 1 }
+      }
+    })
+}
 
-var isDebug = debug.enabled || !!process.env.DEBUG
+if (debug.enabled) {
+  debug('Dat DEBUG mode engaged, enabling quiet mode')
+}
 
 var config = {
   defaults: [
     { name: 'dir', abbr: 'd', help: 'set the directory for Dat' },
     { name: 'logspeed', default: 400 },
-    { name: 'port', default: 3282, help: 'port to use for connections' },
+    { name: 'port', help: 'port to use for connections (default port: 3282 or first available)' },
     { name: 'utp', default: true, boolean: true, help: 'use utp for discovery' },
     { name: 'http', help: 'serve dat over http (default port: 8080)' },
-    { name: 'quiet', default: isDebug, boolean: true }, // neat-log uses quiet for debug right now
+    { name: 'debug', default: !!process.env.DEBUG && !debug.enabled, boolean: true },
+    { name: 'quiet', default: debug.enabled, boolean: true }, // use quiet for dat debugging
     { name: 'sparse', default: false, boolean: true, help: 'download only requested data' },
     { name: 'up', help: 'throttle upload bandwidth (1024, 1kb, 2mb, etc.)' },
     { name: 'down', help: 'throttle download bandwidth (1024, 1kb, 2mb, etc.)' }
@@ -42,11 +60,10 @@ var config = {
   commands: [
     require('../src/commands/clone'),
     require('../src/commands/create'),
-    require('../src/commands/doctor'),
     require('../src/commands/log'),
+    require('../src/commands/keys'),
     require('../src/commands/publish'),
     require('../src/commands/pull'),
-    require('../src/commands/share'),
     require('../src/commands/status'),
     require('../src/commands/sync'),
     require('../src/commands/unpublish'),
@@ -56,20 +73,23 @@ var config = {
     require('../src/commands/auth/login')
   ],
   usage: {
-    command: usage,
+    command: showUsageOrRunExtension,
     option: {
       name: 'help',
       abbr: 'h'
     }
   },
   aliases: {
-    'init': 'create'
+    'init': 'create',
+    'share': 'sync'
   },
-  extensions: [] // whitelist extensions for now
+  // whitelist extensions for now
+  extensions: [
+    'store'
+  ]
 }
 
 if (debug.enabled) {
-  var pkg = require('../package.json')
   debug('dat', pkg.version)
   debug('node', process.version)
 }
@@ -94,7 +114,11 @@ function syncShorthand (opts) {
   if (!opts._.length) return usage(opts)
   debug('Sync shortcut command')
 
-  var parsed = parseArgs(opts)
+  debug('Trying extension', opts._[0])
+  // First try extension
+  if (config.extensions.indexOf(opts._[0]) > -1) return require('../src/extensions')(opts)
+
+  var parsed = require('../src/parse-args')(opts)
 
   // Download Key
   if (parsed.key) {
@@ -114,19 +138,22 @@ function syncShorthand (opts) {
     // Set default opts. TODO: use default opts in share
     opts.watch = opts.watch || true
     opts.import = opts.import || true
-    return require('../src/commands/share').command(opts)
+    return require('../src/commands/sync').command(opts)
   }
-
-  // If directory sync fails, finally try extension
-  if (config.extensions.indexOf(opts._[0]) > -1) return require('../src/extensions')(opts)
 
   // All else fails, show usage
   return usage(opts)
+}
+
+// This was needed so that we can show help messages from extensions
+function showUsageOrRunExtension (opts, help, usageMessage) {
+  if (config.extensions.indexOf(opts._[0]) > -1) return require('../src/extensions')(opts)
+  usage(opts, help, usageMessage)
 }
 
 function exitInvalidNode () {
   console.error('Node Version:', process.version)
   console.error('Unfortunately, we only support Node >= v4. Please upgrade to use Dat.')
   console.error('You can find the latest version at https://nodejs.org/')
-  process.exit(1)
+  process.exit(0)
 }