+++ src/commands/auth/login.js
@@ -1,8 +1,3 @@
-var prompt = require('prompt')
-var output = require('neat-log/output')
-var chalk = require('chalk')
-var Registry = require('../../registry')
-
 module.exports = {
   name: 'login',
   command: login,
@@ -22,19 +17,24 @@ module.exports = {
 }
 
 function login (opts) {
+  var prompt = require('prompt')
+  var output = require('neat-log/output')
+  var chalk = require('chalk')
+  var Registry = require('../../registry')
+
   if (opts._[0]) opts.server = opts._[0]
-  var welcome = output`
+  var welcome = output(`
     Welcome to ${chalk.green(`dat`)} program!
     Login to publish your dats.
 
-  `
+  `)
   console.log(welcome)
 
   var schema = {
     properties: {
       server: {
         description: chalk.magenta('Dat registry'),
-        default: opts.server || 'datproject.org',
+        default: opts.server || 'datbase.org',
         required: true
       },
       email: {
@@ -70,13 +70,13 @@ function login (opts) {
       if (err && err.message) return exitErr(err.message)
       else if (err) return exitErr(err.toString())
 
-      console.log(output`
+      console.log(output(`
         Logged you in to ${chalk.green(opts.server)}!
 
         Now you can publish dats and share:
         * Run ${chalk.green(`dat publish`)} to publish a dat!
         * View & Share your dats at ${opts.server}
-      `)
+      `))
       process.exit(0)
     })
   }