+++ src/commands/clone.js
@@ -1,15 +1,3 @@
-var fs = require('fs')
-var rimraf = require('rimraf')
-var Dat = require('dat-node')
-var linkResolve = require('dat-link-resolve')
-var neatLog = require('neat-log')
-var archiveUI = require('../ui/archive')
-var trackArchive = require('../lib/archive')
-var discoveryExit = require('../lib/discovery-exit')
-var onExit = require('../lib/exit')
-var parseArgs = require('../parse-args')
-var debug = require('debug')('dat')
-
 module.exports = {
   name: 'clone',
   command: clone,
@@ -19,6 +7,12 @@ module.exports = {
     'Usage: dat clone <link> [download-folder]'
   ].join('\n'),
   options: [
+    {
+      name: 'empty',
+      boolean: false,
+      default: false,
+      help: 'Do not download files by default. Files must be synced manually.'
+    },
     {
       name: 'upload',
       boolean: true,
@@ -36,15 +30,39 @@ module.exports = {
 }
 
 function clone (opts) {
+  var fs = require('fs')
+  var path = require('path')
+  var rimraf = require('rimraf')
+  var Dat = require('dat-node')
+  var linkResolve = require('dat-link-resolve')
+  var neatLog = require('neat-log')
+  var archiveUI = require('../ui/archive')
+  var trackArchive = require('../lib/archive')
+  var discoveryExit = require('../lib/discovery-exit')
+  var onExit = require('../lib/exit')
+  var parseArgs = require('../parse-args')
+  var debug = require('debug')('dat')
+
   var parsed = parseArgs(opts)
   opts.key = parsed.key || opts._[0] // pass other links to resolver
   opts.dir = parsed.dir
   opts.showKey = opts['show-key'] // using abbr in option makes printed help confusing
+  opts.sparse = opts.empty
 
   debug('clone()')
-  debug(Object.assign({}, opts, {key: '<private>', _: null})) // don't show key
 
-  var neat = neatLog(archiveUI, { logspeed: opts.logspeed, quiet: opts.quiet })
+  // cmd: dat /path/to/dat.json (opts.key is path to dat.json)
+  if (fs.existsSync(opts.key)) {
+    try {
+      opts.key = getDatJsonKey()
+    } catch (e) {
+      debug('error reading dat.json key', e)
+    }
+  }
+
+  debug(Object.assign({}, opts, { key: '<private>', _: null })) // don't show key
+
+  var neat = neatLog(archiveUI, { logspeed: opts.logspeed, quiet: opts.quiet, debug: opts.debug })
   neat.use(trackArchive)
   neat.use(discoveryExit)
   neat.use(onExit)
@@ -74,6 +92,9 @@ function clone (opts) {
       // Create the directory if it doesn't exist
       // If no dir is specified, we put dat in a dir with name = key
       if (!opts.dir) opts.dir = key
+      if (!Buffer.isBuffer(opts.dir) && typeof opts.dir !== 'string') {
+        return bus.emit('exit:error', 'Directory path must be a string or Buffer')
+      }
       fs.access(opts.dir, fs.F_OK, function (err) {
         if (!err) {
           createdDirectory = false
@@ -101,4 +122,19 @@ function clone (opts) {
       })
     }
   })
+
+  function getDatJsonKey () {
+    var datPath = opts.key
+    var stat = fs.lstatSync(datPath)
+
+    if (stat.isDirectory()) datPath = path.join(datPath, 'dat.json')
+
+    if (!fs.existsSync(datPath) || path.basename(datPath) !== 'dat.json') {
+      if (stat.isFile()) throw new Error('must specify existing dat.json file to read key')
+      throw new Error('directory must contain a dat.json')
+    }
+
+    debug('reading key from dat.json:', datPath)
+    return JSON.parse(fs.readFileSync(datPath, 'utf8')).url
+  }
 }