+++ src/commands/create.js
@@ -1,13 +1,3 @@
-var path = require('path')
-var fs = require('fs')
-var Dat = require('dat-node')
-var output = require('neat-log/output')
-var DatJson = require('dat-json')
-var prompt = require('prompt')
-var chalk = require('chalk')
-var parseArgs = require('../parse-args')
-var debug = require('debug')('dat')
-
 module.exports = {
   name: 'create',
   command: create,
@@ -23,30 +13,48 @@ module.exports = {
       default: false,
       abbr: 'y',
       help: 'Skip dat.json creation.'
+    },
+    {
+      name: 'title',
+      help: 'the title property for dat.json'
+    },
+    {
+      name: 'description',
+      help: 'the description property for dat.json'
     }
   ]
 }
 
 function create (opts) {
+  var path = require('path')
+  var fs = require('fs')
+  var Dat = require('dat-node')
+  var output = require('neat-log/output')
+  var DatJson = require('dat-json')
+  var prompt = require('prompt')
+  var chalk = require('chalk')
+  var parseArgs = require('../parse-args')
+  var debug = require('debug')('dat')
+
   debug('dat create')
   if (!opts.dir) {
     opts.dir = parseArgs(opts).dir || process.cwd()
   }
 
   var welcome = `Welcome to ${chalk.green(`dat`)} program!`
-  var intro = output`
+  var intro = output(`
     You can turn any folder on your computer into a Dat.
     A Dat is a folder with some magic.
 
     Your dat is ready!
-    We will walk you creating a 'dat.json' file.
+    We will walk you through creating a 'dat.json' file.
     (You can skip dat.json and get started now.)
 
     Learn more about dat.json: ${chalk.blue(`https://github.com/datprotocol/dat.json`)}
 
     ${chalk.dim('Ctrl+C to exit at any time')}
 
-  `
+  `)
   var outro
 
   // Force certain options
@@ -57,16 +65,16 @@ function create (opts) {
     if (err && err.name === 'ExistsError') return exitErr('\nArchive already exists.\nYou can use `dat sync` to update.')
     if (err) return exitErr(err)
 
-    outro = output`
+    outro = output(`
 
       Created empty Dat in ${dat.path}/.dat
 
       Now you can add files and share:
       * Run ${chalk.green(`dat share`)} to create metadata and sync.
-      * Copy the unique dat link and securly share it.
+      * Copy the unique dat link and securely share it.
 
       ${chalk.blue(`dat://${dat.key.toString('hex')}`)}
-    `
+    `)
 
     if (opts.yes) return done()