+++ src/commands/publish.js
@@ -1,13 +1,3 @@
-var path = require('path')
-var Dat = require('dat-node')
-var encoding = require('dat-encoding')
-var output = require('neat-log/output')
-var prompt = require('prompt')
-var chalk = require('chalk')
-var DatJson = require('dat-json')
-var xtend = require('xtend')
-var Registry = require('../registry')
-
 module.exports = {
   name: 'publish',
   command: publish,
@@ -27,14 +17,24 @@ module.exports = {
 }
 
 function publish (opts) {
+  var path = require('path')
+  var Dat = require('dat-node')
+  var encoding = require('dat-encoding')
+  var output = require('neat-log/output')
+  var prompt = require('prompt')
+  var chalk = require('chalk')
+  var DatJson = require('dat-json')
+  var xtend = Object.assign
+  var Registry = require('../registry')
+
   if (!opts.dir) opts.dir = process.cwd()
   if (opts._[0]) opts.server = opts._[0]
-  if (!opts.server) opts.server = 'datproject.org' // nicer error message if not logged in
+  if (!opts.server) opts.server = 'datbase.org' // nicer error message if not logged in
 
   var client = Registry(opts)
   var whoami = client.whoami()
   if (!whoami || !whoami.token) {
-    var loginErr = output`
+    var loginErr = output(`
       Welcome to ${chalk.green(`dat`)} program!
       Publish your dats to ${chalk.green(opts.server)}.
 
@@ -44,8 +44,8 @@ function publish (opts) {
       New to ${chalk.green(opts.server)} and need an account?
       ${chalk.green('dat register')}
 
-      Explore public dats at ${chalk.blue('datproject.org/explore')}
-    `
+      Explore public dats at ${chalk.blue('datbase.org/explore')}
+    `)
     return exitErr(loginErr)
   }
 
@@ -56,7 +56,7 @@ function publish (opts) {
 
     dat.joinNetwork() // join network to upload metadata
 
-    var datjson = DatJson(dat.archive, {file: path.join(dat.path, 'dat.json')})
+    var datjson = DatJson(dat.archive, { file: path.join(dat.path, 'dat.json') })
     datjson.read(publish)
 
     function publish (_, data) {
@@ -69,10 +69,10 @@ function publish (opts) {
         title: opts.title,
         description: opts.description
       }, data)
-      var welcome = output`
+      var welcome = output(`
         Publishing dat to ${chalk.green(opts.server)}!
 
-      `
+      `)
       console.log(welcome)
 
       if (datInfo.name) return makeRequest(datInfo)
@@ -101,9 +101,9 @@ function publish (opts) {
         if (err) {
           if (err.message) {
             if (err.message === 'timed out') {
-              return exitErr(output`${chalk.red('\nERROR: ' + opts.server + ' could not connect to your computer.')}
+              return exitErr(output(`${chalk.red('\nERROR: ' + opts.server + ' could not connect to your computer.')}
               Troubleshoot here: ${chalk.green('https://docs.datproject.org/troubleshooting#networking-issues')}
-              `)
+              `))
             }
             var str = err.message.trim()
             if (str === 'jwt expired') return exitErr(`Session expired, please ${chalk.green('dat login')} again`)
@@ -118,26 +118,26 @@ function publish (opts) {
         datjson.write(datInfo, function (err) {
           if (err) return exitErr(err)
           // TODO: write published url to dat.json (need spec)
-          var msg = output`
+          var msg = output(`
 
             We ${body.updated === 1 ? 'updated' : 'published'} your dat!
             ${chalk.blue.underline(`${opts.server}/${whoami.username}/${datInfo.name}`)}
-          `// TODO: get url back? it'd be better to confirm link than guess username/datname structure
+          `)// TODO: get url back? it'd be better to confirm link than guess username/datname structure
 
           console.log(msg)
           if (body.updated === 1) {
-            console.log(output`
+            console.log(output(`
 
               ${chalk.dim.green('Cool fact #21')}
               ${opts.server} will live update when you are sharing your dat!
               You only need to publish again if your dat link changes.
-            `)
+            `))
           } else {
-            console.log(output`
+            console.log(output(`
 
               Remember to use ${chalk.green('dat share')} before sharing.
               This will make sure your dat is available.
-            `)
+            `))
           }
           process.exit(0)
         })