+++ src/commands/status.js
@@ -1,10 +1,3 @@
-var Dat = require('dat-node')
-var neatLog = require('neat-log')
-var statusUI = require('../ui/status')
-var onExit = require('../lib/exit')
-var parseArgs = require('../parse-args')
-var debug = require('debug')('dat')
-
 module.exports = {
   name: 'status',
   command: status,
@@ -17,13 +10,20 @@ module.exports = {
 }
 
 function status (opts) {
+  var Dat = require('dat-node')
+  var neatLog = require('neat-log')
+  var statusUI = require('../ui/status')
+  var onExit = require('../lib/exit')
+  var parseArgs = require('../parse-args')
+  var debug = require('debug')('dat')
+
   debug('dat status')
   if (!opts.dir) {
     opts.dir = parseArgs(opts).dir || process.cwd()
   }
   opts.createIfMissing = false // sync must always be a resumed archive
 
-  var neat = neatLog(statusUI, { logspeed: opts.logspeed, quiet: opts.quiet })
+  var neat = neatLog(statusUI, { logspeed: opts.logspeed, quiet: opts.quiet, debug: opts.debug })
   neat.use(onExit)
   neat.use(function (state, bus) {
     state.opts = opts