+++ src/commands/unpublish.js
@@ -1,11 +1,3 @@
-var prompt = require('prompt')
-var path = require('path')
-var Dat = require('dat-node')
-var output = require('neat-log/output')
-var chalk = require('chalk')
-var DatJson = require('dat-json')
-var Registry = require('../registry')
-
 module.exports = {
   name: 'unpublish',
   command: unpublish,
@@ -25,18 +17,26 @@ module.exports = {
 }
 
 function unpublish (opts) {
+  var prompt = require('prompt')
+  var path = require('path')
+  var Dat = require('dat-node')
+  var output = require('neat-log/output')
+  var chalk = require('chalk')
+  var DatJson = require('dat-json')
+  var Registry = require('../registry')
+
   if (opts._[0]) opts.server = opts._[0]
   if (!opts.dir) opts.dir = process.cwd() // run in dir for `dat unpublish`
 
   var client = Registry(opts)
   var whoami = client.whoami()
   if (!whoami || !whoami.token) {
-    var loginErr = output`
+    var loginErr = output(`
       Welcome to ${chalk.green(`dat`)} program!
 
       ${chalk.bold('You must login before unpublishing.')}
       ${chalk.green('dat login')}
-    `
+    `)
     return exitErr(loginErr)
   }
 
@@ -46,7 +46,7 @@ function unpublish (opts) {
     // TODO better error msg for non-existing archive
     if (!dat.writable) return exitErr('Sorry, you can only publish a dat that you created.')
 
-    var datjson = DatJson(dat.archive, {file: path.join(dat.path, 'dat.json')})
+    var datjson = DatJson(dat.archive, { file: path.join(dat.path, 'dat.json') })
     datjson.read(function (err, data) {
       if (err) return exitErr(err)
       if (!data.name) return exitErr('Try `dat unpublish <name>` with this dat, we are having trouble reading it.')
@@ -73,7 +73,7 @@ function unpublish (opts) {
   }
 
   function makeRequest (name) {
-    client.dats.delete({name: name}, function (err, resp, body) {
+    client.dats.delete({ name: name }, function (err, resp, body) {
       if (err && err.message) exitErr(err.message)
       else if (err) exitErr(err.toString())
       if (body.statusCode === 400) return exitErr(new Error(body.message))