+++ test/clone.js
@@ -1,3 +1,4 @@
+var fs = require('fs')
 var path = require('path')
 var test = require('tape')
 var tempDir = require('temporary-directory')
@@ -11,7 +12,7 @@ test('clone - default opts', function (t) {
     var key = shareDat.key.toString('hex')
     tempDir(function (_, dir, cleanup) {
       var cmd = dat + ' clone ' + key
-      var st = spawn(t, cmd, {cwd: dir})
+      var st = spawn(t, cmd, { cwd: dir })
       var datDir = path.join(dir, key)
 
       st.stdout.match(function (output) {
@@ -79,7 +80,7 @@ test('clone - specify dir', function (t) {
       var key = shareDat.key.toString('hex')
       var customDir = 'my_dir'
       var cmd = dat + ' clone ' + key + ' ' + customDir
-      var st = spawn(t, cmd, {cwd: dir})
+      var st = spawn(t, cmd, { cwd: dir })
       st.stdout.match(function (output) {
         var downloadFinished = output.indexOf('Exiting') > -1
         if (!downloadFinished) return false
@@ -104,7 +105,7 @@ test('clone - dat:// link', function (t) {
       var key = 'dat://' + shareDat.key.toString('hex') + '/'
       var cmd = dat + ' clone ' + key + ' '
       var downloadDir = path.join(dir, shareDat.key.toString('hex'))
-      var st = spawn(t, cmd, {cwd: dir})
+      var st = spawn(t, cmd, { cwd: dir })
       st.stdout.match(function (output) {
         var downloadFinished = output.indexOf('Exiting') > -1
         if (!downloadFinished) return false
@@ -129,7 +130,7 @@ test('clone - datproject.org/key link', function (t) {
       var key = 'datproject.org/' + shareDat.key.toString('hex') + '/'
       var cmd = dat + ' clone ' + key + ' '
       var downloadDir = path.join(dir, shareDat.key.toString('hex'))
-      var st = spawn(t, cmd, {cwd: dir})
+      var st = spawn(t, cmd, { cwd: dir })
       st.stdout.match(function (output) {
         var downloadFinished = output.indexOf('Exiting') > -1
         if (!downloadFinished) return false
@@ -196,7 +197,7 @@ test('clone - invalid link', function (t) {
   var key = 'best-key-ever'
   var cmd = dat + ' clone ' + key
   tempDir(function (_, dir, cleanup) {
-    var st = spawn(t, cmd, {cwd: dir})
+    var st = spawn(t, cmd, { cwd: dir })
     var datDir = path.join(dir, key)
     st.stderr.match(function (output) {
       var error = output.indexOf('Could not resolve link') > -1
@@ -271,3 +272,203 @@ test('clone - shortcut/stateless clone', function (t) {
 //     })
 //   })
 // })
+
+test('clone - specify directory containing dat.json', function (t) {
+  help.shareFixtures(function (_, shareDat) {
+    tempDir(function (_, dir, cleanup) {
+      fs.writeFileSync(path.join(dir, 'dat.json'), JSON.stringify({ url: shareDat.key.toString('hex') }), 'utf8')
+
+      // dat clone /dir
+      var cmd = dat + ' clone ' + dir
+      var st = spawn(t, cmd)
+      var datDir = dir
+
+      st.stdout.match(function (output) {
+        var downloadFinished = output.indexOf('Exiting') > -1
+        if (!downloadFinished) return false
+
+        var fileList = help.fileList(datDir).join(' ')
+        var hasCsvFile = fileList.indexOf('all_hour.csv') > -1
+        t.ok(hasCsvFile, 'csv file downloaded')
+        var hasDatFolder = fileList.indexOf('.dat') > -1
+        t.ok(hasDatFolder, '.dat folder created')
+        var hasSubDir = fileList.indexOf('folder') > -1
+        t.ok(hasSubDir, 'folder created')
+        var hasNestedDir = fileList.indexOf('nested') > -1
+        t.ok(hasNestedDir, 'nested folder created')
+        var hasHelloFile = fileList.indexOf('hello.txt') > -1
+        t.ok(hasHelloFile, 'hello.txt file downloaded')
+
+        st.kill()
+        return true
+      })
+      st.succeeds('exits after finishing download')
+      st.stderr.empty()
+      st.end(function () {
+        cleanup()
+        shareDat.close()
+      })
+    })
+  })
+})
+
+test('clone - specify directory containing dat.json with cwd', function (t) {
+  help.shareFixtures(function (_, shareDat) {
+    tempDir(function (_, dir, cleanup) {
+      fs.writeFileSync(path.join(dir, 'dat.json'), JSON.stringify({ url: shareDat.key.toString('hex') }), 'utf8')
+
+      // cd dir && dat clone /dir/dat.json
+      var cmd = dat + ' clone ' + dir
+      var st = spawn(t, cmd, { cwd: dir })
+      var datDir = dir
+
+      st.stdout.match(function (output) {
+        var downloadFinished = output.indexOf('Exiting') > -1
+        if (!downloadFinished) return false
+
+        var fileList = help.fileList(datDir).join(' ')
+        var hasCsvFile = fileList.indexOf('all_hour.csv') > -1
+        t.ok(hasCsvFile, 'csv file downloaded')
+        var hasDatFolder = fileList.indexOf('.dat') > -1
+        t.ok(hasDatFolder, '.dat folder created')
+        var hasSubDir = fileList.indexOf('folder') > -1
+        t.ok(hasSubDir, 'folder created')
+        var hasNestedDir = fileList.indexOf('nested') > -1
+        t.ok(hasNestedDir, 'nested folder created')
+        var hasHelloFile = fileList.indexOf('hello.txt') > -1
+        t.ok(hasHelloFile, 'hello.txt file downloaded')
+
+        st.kill()
+        return true
+      })
+      st.succeeds('exits after finishing download')
+      st.stderr.empty()
+      st.end(function () {
+        cleanup()
+        shareDat.close()
+      })
+    })
+  })
+})
+
+test('clone - specify dat.json path', function (t) {
+  help.shareFixtures(function (_, shareDat) {
+    tempDir(function (_, dir, cleanup) {
+      var datJsonPath = path.join(dir, 'dat.json')
+      fs.writeFileSync(datJsonPath, JSON.stringify({ url: shareDat.key.toString('hex') }), 'utf8')
+
+      // dat clone /dir/dat.json
+      var cmd = dat + ' clone ' + datJsonPath
+      var st = spawn(t, cmd)
+      var datDir = dir
+
+      st.stdout.match(function (output) {
+        var downloadFinished = output.indexOf('Exiting') > -1
+        if (!downloadFinished) return false
+
+        var fileList = help.fileList(datDir).join(' ')
+        var hasCsvFile = fileList.indexOf('all_hour.csv') > -1
+        t.ok(hasCsvFile, 'csv file downloaded')
+        var hasDatFolder = fileList.indexOf('.dat') > -1
+        t.ok(hasDatFolder, '.dat folder created')
+        var hasSubDir = fileList.indexOf('folder') > -1
+        t.ok(hasSubDir, 'folder created')
+        var hasNestedDir = fileList.indexOf('nested') > -1
+        t.ok(hasNestedDir, 'nested folder created')
+        var hasHelloFile = fileList.indexOf('hello.txt') > -1
+        t.ok(hasHelloFile, 'hello.txt file downloaded')
+
+        st.kill()
+        return true
+      })
+      st.succeeds('exits after finishing download')
+      st.stderr.empty()
+      st.end(function () {
+        cleanup()
+        shareDat.close()
+      })
+    })
+  })
+})
+
+test('clone - specify dat.json path with cwd', function (t) {
+  help.shareFixtures(function (_, shareDat) {
+    tempDir(function (_, dir, cleanup) {
+      var datJsonPath = path.join(dir, 'dat.json')
+      fs.writeFileSync(datJsonPath, JSON.stringify({ url: shareDat.key.toString('hex') }), 'utf8')
+
+      // cd /dir && dat clone /dir/dat.json
+      var cmd = dat + ' clone ' + datJsonPath
+      var st = spawn(t, cmd, { cwd: dir })
+      var datDir = dir
+
+      st.stdout.match(function (output) {
+        var downloadFinished = output.indexOf('Exiting') > -1
+        if (!downloadFinished) return false
+
+        var fileList = help.fileList(datDir).join(' ')
+        var hasCsvFile = fileList.indexOf('all_hour.csv') > -1
+        t.ok(hasCsvFile, 'csv file downloaded')
+        var hasDatFolder = fileList.indexOf('.dat') > -1
+        t.ok(hasDatFolder, '.dat folder created')
+        var hasSubDir = fileList.indexOf('folder') > -1
+        t.ok(hasSubDir, 'folder created')
+        var hasNestedDir = fileList.indexOf('nested') > -1
+        t.ok(hasNestedDir, 'nested folder created')
+        var hasHelloFile = fileList.indexOf('hello.txt') > -1
+        t.ok(hasHelloFile, 'hello.txt file downloaded')
+
+        st.kill()
+        return true
+      })
+      st.succeeds('exits after finishing download')
+      st.stderr.empty()
+      st.end(function () {
+        cleanup()
+        shareDat.close()
+      })
+    })
+  })
+})
+
+test('clone - specify dat.json + directory', function (t) {
+  help.shareFixtures(function (_, shareDat) {
+    tempDir(function (_, dir, cleanup) {
+      var datDir = path.join(dir, 'clone-dest')
+      var datJsonPath = path.join(dir, 'dat.json') // make dat.json in different dir
+
+      fs.mkdirSync(datDir)
+      fs.writeFileSync(datJsonPath, JSON.stringify({ url: shareDat.key.toString('hex') }), 'utf8')
+
+      // dat clone /dir/dat.json /dir/clone-dest
+      var cmd = dat + ' clone ' + datJsonPath + ' ' + datDir
+      var st = spawn(t, cmd)
+
+      st.stdout.match(function (output) {
+        var downloadFinished = output.indexOf('Exiting') > -1
+        if (!downloadFinished) return false
+
+        var fileList = help.fileList(datDir).join(' ')
+        var hasCsvFile = fileList.indexOf('all_hour.csv') > -1
+        t.ok(hasCsvFile, 'csv file downloaded')
+        var hasDatFolder = fileList.indexOf('.dat') > -1
+        t.ok(hasDatFolder, '.dat folder created')
+        var hasSubDir = fileList.indexOf('folder') > -1
+        t.ok(hasSubDir, 'folder created')
+        var hasNestedDir = fileList.indexOf('nested') > -1
+        t.ok(hasNestedDir, 'nested folder created')
+        var hasHelloFile = fileList.indexOf('hello.txt') > -1
+        t.ok(hasHelloFile, 'hello.txt file downloaded')
+
+        st.kill()
+        return true
+      })
+      st.succeeds('exits after finishing download')
+      st.stderr.empty()
+      st.end(function () {
+        cleanup()
+        shareDat.close()
+      })
+    })
+  })
+})